refactor: token으로 사원 번호를 알아내는 코드 미들웨어로 분리 (#104)
